+++ CHANGELOG.md
@@ -2,10 +2,17 @@
 
 ## [Unreleased]
 
+Maintenance update.
+
+- **Changed**: For safety, any drawn text is now drawn with `ImGui.TextUnformatted`.
+- **Changed**: Configuration fields now have consistent indentation.
+- **Changed**: The description for dropdown configuration fields is now correctly below the dropdown.
+
 ### Custom Chat Timestamp
 
 - **Added**: The configuration now shows the result of the current format with an example message.
 - **Added**: The configuration now shows an error if the format is invalid.
+- **Added**: The format field in the configuration now has a reset to default button.
 - **Fixed**: The timestamp will not be modified if the format field is empty.
 
 ## [14.9.1] (2023-04-29)