+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [Unreleased]
+## [14.9.2] (2023-04-30)
 
-Maintenance update.
+Maintenance update with some internal restructuring and bug fixes.
 
 - **Changed**: For safety, any drawn text is now drawn with `ImGui.TextUnformatted`.
 - **Changed**: Configuration fields now have consistent indentation.
@@ -768,6 +768,7 @@ Refreshes the material list and recipe tree when you've crafted or gathered an i
 Opens duty finder for the duty you clicked on in the Wondrous Tails Journal.
 
 [unreleased]: https://github.com/Haselnussbomber/HaselTweaks/compare/main...dev
+[14.9.2]: https://github.com/Haselnussbomber/HaselTweaks/compare/v14.9.1...v14.9.2
 [14.9.1]: https://github.com/Haselnussbomber/HaselTweaks/compare/v14.9.0...v14.9.1
 [14.9.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v14.8.0...v14.9.0
 [14.8.0]: https://github.com/Haselnussbomber/HaselTweaks/compare/v14.7.0...v14.8.0